유저가 참여하고 있는 모든 모임 조회
